requirements.txt
allure-pytest==2.13.2
allure-pytest-bdd==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==2.11.1
assertpy==1.1
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
exceptiongroup==1.1.1
Faker==18.10.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
jsondiff==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
outcome==1.2.0
packaging==23.1
parse==1.19.0
parse-type==0.6.0
playwright==1.33.0
pluggy==1.0.0
pycparser==2.21
pyee==9.0.4
PySocks==1.7.1
pytest==7.3.1
pytest-base-url==2.0.0
pytest-bdd==6.1.1
pytest-playwright==0.3.3
python-dateutil==2.8.2
python-slugify==8.0.1
requests==2.31.0
screeninfo==0.8.1
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sttable==0.0.1
text-unidecode==1.3
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.2
urllib3==2.0.2
wsproto==1.2.0
